refactor(data): move some mnemonic schemas and data validation logic to a public module
